Added the register page
Added a page to register to Beacon, as well as removing some unnecessary code. Also figured out that the way @Susorodni was getting the value of inputs was incorrect, will be fixing it in my next commit. Also added package-lock to the gitignore, we usin' yarn.

Also removed thing because why  is that there lol.